+++ chatlas/_openai.py
@@ -6,7 +6,15 @@ from typing import TYPE_CHECKING, Any, Literal, Optional, cast, overload
 from pydantic import BaseModel
 
 from ._chat import Chat
-from ._content import (
+from ._merge import merge_dicts
+from ._provider import Provider
+from ._tokens import tokens_log
+from ._tools import Tool, ToolSchema, basemodel_to_param_schema
+from ._turn import Turn, normalize_turns
+from ._utils import inform_model_default, is_testing
+from .types import (
+    MISSING,
+    MISSING_TYPE,
     Content,
     ContentImageInline,
     ContentImageRemote,
@@ -15,12 +23,6 @@ from ._content import (
     ContentToolRequest,
     ContentToolResult,
 )
-from ._merge import merge_dicts
-from ._provider import Provider
-from ._tokens import tokens_log
-from ._tools import Tool, ToolSchema, basemodel_to_param_schema
-from ._turn import Turn, normalize_turns
-from ._utils import MISSING, MISSING_TYPE, inform_model_default, is_testing
 
 if TYPE_CHECKING:
     from openai.types.chat import (
@@ -37,9 +39,11 @@ if TYPE_CHECKING:
     )
     from openai.types.chat_model import ChatModel
 
-    from .types._openai_client import ProviderClientArgs
-    from .types._openai_client_azure import ProviderClientArgs as AzureProviderArgs
-    from .types._openai_create import ChatCompletionArgs
+    from .provider_types._openai_client import ProviderClientArgs
+    from .provider_types._openai_client_azure import (
+        ProviderClientArgs as AzureProviderArgs,
+    )
+    from .provider_types._openai_create import ChatCompletionArgs
 else:
     ChatCompletion = object
     ChatCompletionChunk = object
