+++ chatlas/_turn.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any, Optional, Sequence
 
-from ._content import Content, ContentText
+from .types import Content, ContentText
 
 __all__ = ("Turn",)
