+++ docs/reference/ChatAzureOpenAI.qmd
@@ -29,7 +29,7 @@ from OpenAI.
 | api_key       | [Optional](`typing.Optional`)\[[str](`str`)\]                                   | The API key to use for authentication. You generally should not supply this directly, but instead set the `AZURE_OPENAI_API_KEY` environment variable.                                                                                                                                                                                                                                                                    | `None`     |
 | system_prompt | [Optional](`typing.Optional`)\[[str](`str`)\]                                   | A system prompt to set the behavior of the assistant.                                                                                                                                                                                                                                                                                                                                                                     | `None`     |
 | turns         | [Optional](`typing.Optional`)\[[list](`list`)\[[Turn](`chatlas._turn.Turn`)\]\] | A list of turns to start the chat with (i.e., continuing a previous conversation). If not provided, the conversation begins from scratch. Do not provide non-None values for both `turns` and `system_prompt`. Each message in the list should be a dictionary with at least `role` (usually `system`, `user`, or `assistant`, but `tool` is also possible). Normally there is also a `content` field, which is a string. | `None`     |
-| seed          | [int](`int`) \| None \| [MISSING_TYPE](`chatlas._utils.MISSING_TYPE`)           | Optional integer seed that ChatGPT uses to try and make output more reproducible.                                                                                                                                                                                                                                                                                                                                         | `MISSING`  |
+| seed          | [int](`int`) \| None \| [MISSING_TYPE](`chatlas.types.MISSING_TYPE`)            | Optional integer seed that ChatGPT uses to try and make output more reproducible.                                                                                                                                                                                                                                                                                                                                         | `MISSING`  |
 | kwargs        | [Optional](`typing.Optional`)\[\'AzureProviderArgs\'\]                          | Additional arguments to pass to the `openai.AzureOpenAI()` client constructor.                                                                                                                                                                                                                                                                                                                                            | `None`     |
 
 ## Returns {.doc-section .doc-section-returns}
