+++ docs/reference/ChatOpenAI.qmd
@@ -31,7 +31,7 @@ that will give you an API key that you can use with this package.
 | model         | \'Optional\[ChatModel \| str\]\'                                                | The model to use for the chat. The default, None, will pick a reasonable default, and warn you about it. We strongly recommend explicitly choosing a model for all but the most casual use.                                                                                                                                                                                                                                 | `None`                        |
 | api_key       | [Optional](`typing.Optional`)\[[str](`str`)\]                                   | The API key to use for authentication. You generally should not supply this directly, but instead set the `OPENAI_API_KEY` environment variable.                                                                                                                                                                                                                                                                            | `None`                        |
 | base_url      | [str](`str`)                                                                    | The base URL to the endpoint; the default uses OpenAI.                                                                                                                                                                                                                                                                                                                                                                      | `'https://api.openai.com/v1'` |
-| seed          | [int](`int`) \| None \| [MISSING_TYPE](`chatlas._utils.MISSING_TYPE`)           | Optional integer seed that ChatGPT uses to try and make output more reproducible.                                                                                                                                                                                                                                                                                                                                           | `MISSING`                     |
+| seed          | [int](`int`) \| None \| [MISSING_TYPE](`chatlas.types.MISSING_TYPE`)            | Optional integer seed that ChatGPT uses to try and make output more reproducible.                                                                                                                                                                                                                                                                                                                                           | `MISSING`                     |
 | kwargs        | [Optional](`typing.Optional`)\[\'ProviderClientArgs\'\]                         | Additional arguments to pass to the `openai.OpenAI()` client constructor.                                                                                                                                                                                                                                                                                                                                                   | `None`                        |
 
 ## Returns {.doc-section .doc-section-returns}
