+++ docs/reference/content_image_file.qmd
@@ -14,7 +14,7 @@ can handle various image formats and provides options for resizing.
 | Name         | Type                                                                                                | Description                                                                                                                                                                                                                 | Default    |
 |--------------|-----------------------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|------------|
 | path         | [str](`str`)                                                                                        | The path to the image file to include in the chat input.                                                                                                                                                                    | _required_ |
-| content_type | [Literal](`typing.Literal`)\[\'auto\', [ImageContentTypes](`chatlas._content.ImageContentTypes`)\]  | The content type of the image (e.g., "image/png"). If "auto", the content type is inferred from the file extension.                                                                                                         | `'auto'`   |
+| content_type | [Literal](`typing.Literal`)\[\'auto\', [ImageContentTypes](`chatlas.types.ImageContentTypes`)\]     | The content type of the image (e.g., "image/png"). If "auto", the content type is inferred from the file extension.                                                                                                         | `'auto'`   |
 | resize       | [Union](`typing.Union`)\[[str](`str`), [Literal](`typing.Literal`)\[\'none\', \'low\', \'high\'\]\] | Resizing option for the image. Can be:     - "none": No resizing     - "low": Resize to fit within 512x512     - "high": Resize to fit within 2000x768 or 768x2000     - Custom string (e.g., "200x200", "300x200>!", etc.) | `'low'`    |
 
 ## Returns {.doc-section .doc-section-returns}
